mv event types to `EventTypeDefs`
