+++ src/typedefs.ts
@@ -1,10 +1,7 @@
 // https://www.typescriptlang.org/docs/handbook/utility-types.html
-import type { Control } from './controls/control.class';
 import type { Gradient } from './gradient/gradient.class';
 import type { Pattern } from './pattern.class';
 import type { Point } from './point.class';
-import type { FabricObject } from './shapes/fabricObject.class';
-import type { saveObjectTransform } from './util/misc/objectTransforms';
 
 interface NominalTag<T> {
   nominalTag?: T;
@@ -75,66 +72,6 @@ export type TMat2D = [number, number, number, number, number, number];
  * SVG path commands
  */
 export type PathData = (string | number)[][];
-
-export type ModifierKey = 'altKey' | 'shiftKey' | 'ctrlKey';
-
-export type TPointerEvent = MouseEvent | TouchEvent;
-
-export type TransformAction<T extends Transform = Transform, R = void> = (
-  eventData: TPointerEvent,
-  transform: T,
-  x: number,
-  y: number
-) => R;
-
-export type TransformActionHandler<T extends Transform = Transform> =
-  TransformAction<T, boolean>;
-
-export type ControlCallback<R = void> = (
-  eventData: TPointerEvent,
-  control: Control,
-  fabricObject: FabricObject
-) => R;
-
-export type ControlCursorCallback = ControlCallback<string>;
-
-/**
- * relative to target's containing coordinate plane
- * both agree on every point
- */
-export type Transform = {
-  target: FabricObject;
-  action: string;
-  actionHandler: TransformActionHandler;
-  corner: string;
-  scaleX: number;
-  scaleY: number;
-  skewX: number;
-  skewY: number;
-  offsetX: number;
-  offsetY: number;
-  originX: TOriginX;
-  originY: TOriginY;
-  ex: number;
-  ey: number;
-  lastX: number;
-  lastY: number;
-  theta: TRadian;
-  width: number;
-  height: number;
-  shiftKey: boolean;
-  altKey: boolean;
-  original: ReturnType<typeof saveObjectTransform>;
-};
-
-export type TEvent<E extends Event = TPointerEvent> = {
-  e: E;
-};
-
-export type TransformEvent<E extends Event = TPointerEvent> = TEvent<E> & {
-  transform: Transform;
-  pointer: Point;
-};
 
 /**
  * An invalid keyword and an empty string will be handled as the `anonymous` keyword.
